Tabular: Move deprecated methods to mixin (#3795)
